+++ tests_requirements/test_pack_1/test_model.py
@@ -1,0 +1,10 @@
+from test_pack_2 import name
+
+
+class TestM:
+
+    name = name
+
+    def _init_(self, alpha: float, max_lag: int):
+        self.alpha = alpha
+        self.max_lag = max_lag